```go
// replace
func matchSelectorKeys(rv any, keys []string) []string

// with
func matchSelectorKeys(selector *ResolvedValue, keys []string) []string
```

```console
$ benchstat a.txt b.txt
goos: darwin
goarch: arm64
pkg: go.expect.digital/mf2/template
cpu: Apple M2 Max
                   │    a.txt    │               b.txt                │
                   │   sec/op    │   sec/op     vs base               │
Template_Sprint-12   2.477µ ± 1%   2.433µ ± 1%  -1.78% (p=0.001 n=10)

                   │    a.txt     │                b.txt                │
                   │     B/op     │     B/op      vs base               │
Template_Sprint-12   4.391Ki ± 0%   4.297Ki ± 0%  -2.14% (p=0.000 n=10)

                   │   a.txt    │               b.txt               │
                   │ allocs/op  │ allocs/op   vs base               │
Template_Sprint-12   88.00 ± 0%   86.00 ± 0%  -2.27% (p=0.000 n=10)
```